almost done

+++ Tar5/symbol_table_records.zig
@@ -2,11 +2,11 @@ const std = @import("std");
 
 // struct of the Class table record
 pub const ClassRecord = struct {
-    var_type: []const u8,
+    var_type: []const usize,
     static: bool,
     index: usize,
 
-    pub fn init(var_type: []const u8, static: bool, index: u32) ClassRecord {
+    pub fn init(var_type: []const u8, static: bool, index: usize) ClassRecord {
         return ClassRecord{
             .var_type = var_type,
             .static = static,
@@ -21,11 +21,11 @@ pub const ClassRecord = struct {
 
 // struct of the Function table record
 pub const FunctionRecord = struct {
-    return_type: []const u8,
+    return_type: []const usize,
     argument: bool,
     index: usize,
 
-    pub fn init(return_type: []const u8, argument: bool, index: u32) FunctionRecord {
+    pub fn init(return_type: []const u8, argument: bool, index: usize) FunctionRecord {
         return FunctionRecord{
             .return_type = return_type,
             .argument = argument,
